requirements.txt
django>=4.0,<5.0
djangorestframework>=3.13,<4.0
psycopg2-binary>=2.9
